Package cleanup, don't clobber tmux buffer with yankee

+++ hooks/toolset/neovim.sh
@@ -1,11 +1,16 @@
 toolset_subsection "Neovim"
 
 if ! command -v nvim >/dev/null; then
-  neovim_appimage=$(curl "https://api.github.com/repos/neovim/neovim/releases/latest" \
-    | jq -r '.assets[] | select(.name == "nvim.appimage") | .browser_download_url')
+  if uname | grep -F Linux | grep -Fvq Microsoft && command -v jq >/dev/null; then
+    neovim_appimage=$(curl "https://api.github.com/repos/neovim/neovim/releases/latest" \
+      | jq -r '.assets[] | select(.name == "nvim.appimage") | .browser_download_url')
 
-  info "Downloading latest NeoVim AppImage..."
-  curl -L "$neovim_appimage" -o "$HOME/.local/bin/nvim"
+    info "Neovim is unavailable! Installing from AppImage..."
+    curl -L "$neovim_appimage" -o "$HOME/.local/bin/nvim"
+  else
+    important "Neovim is unavailable!"
+    return
+  fi
 fi
 
 if has_support "Python"; then
@@ -17,7 +22,7 @@ if has_support "Python"; then
       info "Creating $venv ($vers) virtual environment..."
       pyenv uninstall -f "$venv"
       pyenv virtualenv "$vers" "$venv"
-      PYENV_VERSION=$vers pip install pynvim
+      PYENV_VERSION=$venv pip install pynvim
     fi
   done
 fi
